+++ osquery/database/database.cpp
@@ -248,7 +248,7 @@ inline void addLegacyFieldsAndDecorations(const QueryLogItem& item,
   tree.put<std::string>("name", item.name);
   tree.put<std::string>("hostIdentifier", item.identifier);
   tree.put<std::string>("calendarTime", item.calendar_time);
-  tree.put<int>("unixTime", item.time);
+  tree.put<size_t>("unixTime", item.time);
 
   // Append the decorations.
   if (item.decorations.size() > 0) {
